# specs/001-gitlab-repository-mining/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Repository Miner API
  version: 0.3.0
  description: >-
    Durable REST contract for incremental GitLab commit-author monitoring. Operator identity is
    validated by the hosting environment; this API does not provide operator registration, login,
    password, or account-management operations.
servers:
  - url: /api/v1
paths:
  /configurations:
    get:
      operationId: listConfigurations
      responses:
        "200":
          description: Configurations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigurationPage" }
    post:
      operationId: createConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigurationWrite" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Configuration" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /configurations/{configurationId}:
    parameters:
      - $ref: "#/components/parameters/ConfigurationId"
    get:
      operationId: getConfiguration
      responses:
        "200":
          description: Configuration without credential value
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Configuration" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigurationWrite" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Configuration" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /configurations/{configurationId}/repository-selections:
    parameters:
      - $ref: "#/components/parameters/ConfigurationId"
    put:
      operationId: replaceRepositorySelections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rules]
              properties:
                target_branch: { type: string, minLength: 1, default: QA }
                rules:
                  type: array
                  items: { $ref: "#/components/schemas/SelectionRule" }
      responses:
        "200": { description: Saved branch and repository selection after successful connection validation }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /configurations/{configurationId}/allowed-users:
    parameters:
      - $ref: "#/components/parameters/ConfigurationId"
    put:
      operationId: replaceAllowedUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [users]
              properties:
                users:
                  type: array
                  minItems: 1
                  items: { $ref: "#/components/schemas/AllowedUser" }
      responses:
        "200": { description: Saved normalized allowed-user list }
        "422": { $ref: "#/components/responses/ValidationError" }
  /configurations/{configurationId}/credential-incidents:
    parameters:
      - $ref: "#/components/parameters/ConfigurationId"
    post:
      operationId: markGitLabCredentialCompromised
      description: >-
        Requires a valid host-authenticated operator. The validated operator identity is obtained
        by the HTTP adapter and recorded in the secret-free incident audit; it is not accepted in
        the request body.
      security:
        - hostAuthentication: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [safe_reason]
              properties:
                safe_reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "201":
          description: Credential blocked and incident recorded without the secret
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialIncident" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /configurations/{configurationId}/credential-replacements:
    parameters:
      - $ref: "#/components/parameters/ConfigurationId"
    post:
      operationId: replaceGitLabCredential
      description: >-
        Requires a valid host-authenticated operator. The validated operator identity is recorded
        in the replacement audit and is not accepted in the request body. An active credential may
        be replaced preventively; a compromised credential must be replaced before further GitLab
        use. Replacement preserves checkpoints, executions, verifications, and alerts, updates the
        configuration to the new active credential, and invalidates connection validation until a
        successful connection test.
      security:
        - hostAuthentication: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gitlab_token]
              properties:
                gitlab_token: { type: string, writeOnly: true, minLength: 1 }
      responses:
        "200":
          description: New credential stored and replacement audit recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CredentialReplacement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    hostAuthentication:
      type: http
      scheme: bearer
      bearerFormat: host-issued credential
      description: >-
        Transport representation of an identity validated by the hosting environment. The backend
        adapter maps it to AuthenticatedOperator; application use cases do not depend on this bearer
        representation, and deployments may replace the adapter and concrete mechanism.
  parameters:
    ConfigurationId:
      name: configurationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State conflict, including active execution or compromised credential
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Operator lacks resource authorization
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        trace_id: { type: string }
...
    ConfigurationWrite:
      type: object
      required: [name, gitlab_base_url, timezone, schedule]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        gitlab_base_url: { type: string, format: uri }
        gitlab_token: { type: [string, "null"], writeOnly: true }
        timezone: { type: string }
        enabled: { type: boolean, default: false }
        schedule: { $ref: "#/components/schemas/Schedule" }
    Configuration:
      type: object
      required: [id, name, gitlab_base_url, timezone, enabled, schedule, credential_status, connection_validated]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        gitlab_base_url: { type: string, format: uri }
        target_branch: { type: [string, "null"], description: Persisted with repository scope only after successful connection validation; defaults to QA on first save }
        timezone: { type: string }
        enabled: { type: boolean }
        schedule: { $ref: "#/components/schemas/Schedule" }
        credential_status: { type: string, enum: [active, compromised], description: Status of the credential currently linked to the configuration; replaced is historical only }
        connection_validated: { type: boolean, description: True only when validation matches the current GitLab URL and credential reference }
    ConfigurationPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Configuration" }
        next_cursor: { type: [string, "null"] }
    Schedule:
      type: object
      required: [recurrence, local_time]
      properties:
        recurrence: { type: string, enum: [daily, weekly, monthly] }
        local_time: { type: string, format: time }
        weekday: { type: [integer, "null"], minimum: 1, maximum: 7 }
        day_of_month: { type: [integer, "null"], minimum: 1, maximum: 31, description: Missing days use the month's last calendar day }
...
    SelectionRule:
      type: object
      required: [kind, external_id, path, mode]
      properties:
        kind: { type: string, enum: [group, subgroup, repository] }
        external_id: { type: string }
        path: { type: string }
        mode: { type: string, enum: [include, exclude] }
    AllowedUser:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, description: Compared after trim and case normalization }
    CredentialIncident:
      type: object
      required: [id, credential_status, incident_status, suspected_at, suspected_by, safe_reason, operator_action]
      properties:
        id: { type: string, format: uuid }
        credential_status: { type: string, enum: [compromised, replaced] }
        incident_status: { type: string, enum: [open, replacement_recorded, closed] }
        suspected_at: { type: string, format: date-time }
        suspected_by: { type: string, description: Stable host-authenticated operator identifier }
        safe_reason: { type: string }
        operator_action: { type: string, description: Instructs external invalidation or rotation }
        replacement_recorded_at: { type: [string, "null"], format: date-time }
        replacement_recorded_by: { type: [string, "null"], description: Stable host-authenticated operator identifier }
    CredentialReplacement:
      type: object
      required: [id, configuration_id, previous_credential_id, replacement_credential_id, reason, replaced_by, replaced_at]
      properties:
        id: { type: string, format: uuid }
        configuration_id: { type: string, format: uuid }
        previous_credential_id: { type: string, format: uuid }
        replacement_credential_id: { type: string, format: uuid }
        reason: { type: string, enum: [preventive, compromise_remediation] }
        incident_id: { type: [string, "null"], format: uuid }
        replaced_by: { type: string, description: Stable host-authenticated operator identifier }
        replaced_at: { type: string, format: date-time }
...